// front/src/main/web/src/app/api/model/scheduledTaskExecutionDto.ts
/**
 * i18n Tool
 * Web API of the i18n tool
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LocalizedStringDto } from './localizedStringDto';

/**
 * Execution of a scheduled task.
 */
export interface ScheduledTaskExecutionDto { 
    /**
     * The unique id of this execution.
     */
    id: string;
    /**
     * The associated task definition
     */
    definitionId: string;
    /**
     * Time when the execution started.
     */
    startTime: Date;
    /**
     * Time when the execution ended.
     */
    endTime: Date;
    /**
     * Status at the end of the execution.
     */
    status: ScheduledTaskExecutionDto.StatusDtoEnum;
    /**
     * Short description (characters are truncated if there is not enough place on the screen) of the result to be displayed to the end-user. The description may contain HTML.
     */
    shortDescription: LocalizedStringDto;
    /**
     * Description of the result to be displayed to the end-user. If no description is provided, the short description is used. The description may contain HTML.
     */
    description: LocalizedStringDto;
    /**
     * The duration in milli-seconds of this execution.
     */
    durationInMs: number;
}
export namespace ScheduledTaskExecutionDto {
    export type StatusDtoEnum = 'SUCCESSFUL' | 'FAILED';
    export const StatusDtoEnum = {
        SUCCESSFUL: 'SUCCESSFUL' as StatusDtoEnum,
        FAILED: 'FAILED' as StatusDtoEnum
    };
}